Fix code style and fix health check for prefill
