+++ python/sglang/srt/utils.py
@@ -44,7 +44,6 @@ import traceback
 import warnings
 from collections import OrderedDict, defaultdict
 from contextlib import contextmanager
-from enum import Enum
 from functools import lru_cache
 from importlib.metadata import PackageNotFoundError, version
 from importlib.util import find_spec
@@ -92,19 +91,6 @@ logger = logging.getLogger(__name__)
 
 show_time_cost = False
 time_infos = {}
-
-
-#########################
-# Constants & Enums
-#########################
-class ServerStatus(Enum):
-    Up = "Up"
-    Starting = "Starting"
-    UnHealthy = "UnHealthy"
-    Crashed = "Crashed"
-
-    def is_healthy(self) -> bool:
-        return self == ServerStatus.Up
 
 
 HIP_FP8_E4M3_FNUZ_MAX = 224.0
